update compass-instance to shared config
