--- packages/compass-instance/.browserslistrc
@@ -1,1 +1,0 @@
-Electron >= 6
