--- packages/compass-instance/.editorconfig
@@ -1,16 +1,0 @@
-# EditorConfig helps developers define and maintain
-# consistent coding styles between different editors and IDEs.
-
-root = true
-
-[*]
-charset = utf-8
-indent_style = space
-indent_size = 2
-end_of_line = lf
-insert_final_newline = true
-trim_trailing_whitespace = true
-
-[*.md]
-insert_final_newline = true
-trim_trailing_whitespace = true
