+++ packages/compass-instance/.eslintignore
@@ -1,0 +1,2 @@
+.nyc-output
+dist
