+++ packages/compass-instance/.eslintrc.js
@@ -1,0 +1,12 @@
+module.exports = {
+  root: true,
+  extends: ['@mongodb-js/eslint-config-compass'],
+  parserOptions: {
+    tsconfigRootDir: __dirname,
+    project: ['./tsconfig-lint.json'],
+  },
+  env: {
+    node: true,
+    browser: true,
+  },
+};
