--- packages/compass-instance/.gitignore
@@ -1,13 +1,0 @@
-.DS_Store
-npm-debug.log
-node_modules/
-lib/
-lib-cov
-coverage
-.lock-wscript
-.idea/
-*.iml
-*.swp
-.nvmrc
-.nyc_output
-src/assets/fonts/akzid*
