--- packages/compass-instance/.npmignore
@@ -1,19 +1,0 @@
-.babelrc
-.eslintrc
-.editorconfig
-.nycrc
-.travis.yml
-config/
-coverage/
-electron/
-.github/
-karma.conf.js
-lib/index.html
-mocha-webpack.opts
-scripts/
-test/
-src/
-.nyc_output/
-
-scripts/**/*
-!scripts/download-fonts.js
