--- packages/compass-instance/.nycrc
@@ -1,13 +1,0 @@
-{
-  "exclude": [
-    "config",
-    ".tmp",
-    "test"
-  ],
-  "reporter": [
-    "text",
-    "html"
-  ],
-  "sourceMap": false,
-  "instrument": false
-}
