+++ packages/compass-instance/.prettierignore
@@ -1,0 +1,3 @@
+.nyc_output
+dist
+coverage
