--- packages/compass-instance/config/webpack.karma.config.js
@@ -1,29 +1,0 @@
-const merge = require('webpack-merge');
-const baseWebpackConfig = require('./webpack.base.config');
-
-const config = {
-  target: 'electron-renderer', // webpack should compile node compatible code for tests
-  devtool: 'eval-source-map',
-  stats: 'errors-only',
-  externals: {
-    'jsdom': 'window',
-    'react/addons': 'react',
-    'react/lib/ExecutionEnvironment': 'react',
-    'react/lib/ReactContext': 'react',
-    'react-addons-test-utils': 'react-dom'
-  },
-  module: {
-    rules: [
-      {
-        test: /\.(png|jpg|jpeg|gif|svg)$/,
-        use: [{ loader: 'ignore-loader' }]
-      },
-      {
-        test: /\.(woff|woff2|ttf|eot)(\?v=\d+\.\d+\.\d+)?$/,
-        use: [{ loader: 'ignore-loader' }]
-      }
-    ]
-  }
-};
-
-module.exports = merge.smart(baseWebpackConfig, config);
