--- packages/compass-instance/src/components/instance.module.less
@@ -1,3 +1,0 @@
-.instance-error {
-  padding: 16px;
-}
