+++ packages/compass-instance/src/typings.d.ts
@@ -1,0 +1,1 @@
+declare module '@mongodb-js/mongodb-redux-common/app-registry';
