--- packages/compass-instance/test/renderer/store.spec.js
@@ -1,5 +1,0 @@
-describe('InstanceStore [Store]', () => {
-  it('placeholder test', () => {
-    expect(true).to.equal(true);
-  });
-});
