--- packages/compass-instance/test/setup.js
@@ -1,35 +1,0 @@
-const Enzyme = require('enzyme');
-const Adapter = require('enzyme-adapter-react-16');
-Enzyme.configure({ adapter: new Adapter() });
-
-const chai = require('chai');
-const chaiEnzyme = require('chai-enzyme');
-const sinonChai = require('sinon-chai');
-const app = require('hadron-app');
-
-require('jsdom-global')('', {
-  beforeParse(win) {
-    win.URL = {
-      createObjectURL: () => {}
-    };
-  }
-});
-
-global.sinon = require('sinon');
-global.expect = chai.expect;
-global.hadronApp = app;
-
-class Worker {
-  postMessage() {
-    return true;
-  }
-  terminate() {
-    return true;
-  }
-}
-
-global.Worker = Worker;
-
-chai.should();
-chai.use(sinonChai);
-chai.use(chaiEnzyme());
