+++ packages/compass-schema/src/index.ts
@@ -19,7 +19,6 @@ const ROLE = {
 
 /**
  * Activate all the components in the Compass Schema package.
- * @param {Object} appRegistry - The Hadron appRegisrty to activate this plugin with.
  **/
 function activate(appRegistry: AppRegistry): void {
   appRegistry.registerRole('Collection.Tab', ROLE);
@@ -27,7 +26,6 @@ function activate(appRegistry: AppRegistry): void {
 
 /**
  * Deactivate all the components in the Compass Schema package.
- * @param {Object} appRegistry - The Hadron appRegisrty to deactivate this plugin with.
  **/
 function deactivate(appRegistry: AppRegistry): void {
   appRegistry.deregisterRole('Collection.Tab', ROLE);
